Extract DiagMessage from Result
